+++ src/main/scala/com/databricks/labs/automl/model/GBTreesTuner.scala
@@ -459,7 +459,8 @@ class GBTreesTuner(df: DataFrame, modelSelection: String)
     var bestScore: Double = 0.0
     var rollingImprovement: Boolean = true
     var incrementalImprovementCount: Int = 0
-    val earlyStoppingImprovementThreshold: Int = -10
+    val earlyStoppingImprovementThreshold: Int =
+      _continuousEvolutionImprovementThreshold
 
     // Generate the first pool of attempts to seed the hyperparameter space
     //    var runSet = ParHashSet(generateThresholdedParams(_firstGenerationGenePool): _*)
@@ -665,7 +666,7 @@ class GBTreesTuner(df: DataFrame, modelSelection: String)
 
           val expandedCandidates = irradiateGeneration(
             generateIdealParents(currentState),
-            _numberOfMutationsPerGeneration * 25,
+            _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
             mutationAggressiveness,
             _geneticMixing
           )
@@ -673,7 +674,7 @@ class GBTreesTuner(df: DataFrame, modelSelection: String)
           val evolution = GenerationOptimizer
             .gbtCandidates(
               "GBT",
-              "XGBoost",
+              _geneticMBORegressorType,
               fossilRecord,
               expandedCandidates,
               _optimizationStrategy,
@@ -711,7 +712,7 @@ class GBTreesTuner(df: DataFrame, modelSelection: String)
 
         val expandedCandidates = irradiateGeneration(
           generateIdealParents(currentState),
-          _numberOfMutationsPerGeneration * 25,
+          _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
           mutationAggressiveness,
           _geneticMixing
         )
@@ -719,7 +720,7 @@ class GBTreesTuner(df: DataFrame, modelSelection: String)
         val evolution = GenerationOptimizer
           .gbtCandidates(
             "GBT",
-            "XGBoost",
+            _geneticMBORegressorType,
             fossilRecord,
             expandedCandidates,
             _optimizationStrategy,