+++ src/main/scala/com/databricks/labs/automl/model/MLPCTuner.scala
@@ -358,7 +358,8 @@ class MLPCTuner(df: DataFrame)
     var rollingImprovement: Boolean = true
     var incrementalImprovementCount: Int = 0
 
-    val earlyStoppingImprovementThreshold: Int = -10
+    val earlyStoppingImprovementThreshold: Int =
+      _continuousEvolutionImprovementThreshold
 
     val totalConfigs = modelConfigLength[MLPCConfig]
 
@@ -575,7 +576,7 @@ class MLPCTuner(df: DataFrame)
 
           val expandedCandidates = irradiateGeneration(
             generateIdealParents(currentState),
-            _numberOfMutationsPerGeneration * 25,
+            _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
             mutationAggressiveness,
             _geneticMixing
           )
@@ -583,7 +584,7 @@ class MLPCTuner(df: DataFrame)
           val evolution = GenerationOptimizer
             .mlpcCandidates(
               "MLPC",
-              "XGBoost",
+              _geneticMBORegressorType,
               fossilRecord,
               expandedCandidates,
               _optimizationStrategy,
@@ -623,7 +624,7 @@ class MLPCTuner(df: DataFrame)
 
         val expandedCandidates = irradiateGeneration(
           generateIdealParents(currentState),
-          _numberOfMutationsPerGeneration * 25,
+          _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
           mutationAggressiveness,
           _geneticMixing
         )
@@ -631,7 +632,7 @@ class MLPCTuner(df: DataFrame)
         val evolution = GenerationOptimizer
           .mlpcCandidates(
             "MLPC",
-            "XGBoost",
+            _geneticMBORegressorType,
             fossilRecord,
             expandedCandidates,
             _optimizationStrategy,