+++ src/main/scala/com/databricks/labs/automl/model/SVMTuner.scala
@@ -311,7 +311,8 @@ class SVMTuner(df: DataFrame)
     var bestScore: Double = 0.0
     var rollingImprovement: Boolean = true
     var incrementalImprovementCount: Int = 0
-    val earlyStoppingImprovementThreshold: Int = -10
+    val earlyStoppingImprovementThreshold: Int =
+      _continuousEvolutionImprovementThreshold
 
     val totalConfigs = modelConfigLength[SVMConfig]
 
@@ -512,7 +513,7 @@ class SVMTuner(df: DataFrame)
 
           val expandedCandidates = irradiateGeneration(
             generateIdealParents(currentState),
-            _numberOfMutationsPerGeneration * 25,
+            _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
             mutationAggressiveness,
             _geneticMixing
           )
@@ -520,7 +521,7 @@ class SVMTuner(df: DataFrame)
           val evolution = GenerationOptimizer
             .svmCandidates(
               "SVM",
-              "XGBoost",
+              _geneticMBORegressorType,
               fossilRecord,
               expandedCandidates,
               _optimizationStrategy,
@@ -558,7 +559,7 @@ class SVMTuner(df: DataFrame)
 
         val expandedCandidates = irradiateGeneration(
           generateIdealParents(currentState),
-          _numberOfMutationsPerGeneration * 25,
+          _numberOfMutationsPerGeneration * _geneticMBOCandidateFactor,
           mutationAggressiveness,
           _geneticMixing
         )
@@ -566,7 +567,7 @@ class SVMTuner(df: DataFrame)
         val evolution = GenerationOptimizer
           .svmCandidates(
             "SVM",
-            "XGBoost",
+            _geneticMBORegressorType,
             fossilRecord,
             expandedCandidates,
             _optimizationStrategy,